Movie Dist Import Update 2

+++ database/migrations/2020_12_04_223719_create_producers_table.php
@@ -21,7 +21,7 @@ class CreateProducersTable extends Migration
             $table->string('city')->nullable();
             $table->string('country')->nullable();
             $table->string('language')->nullable();
-            $table->integer('share')->nullable();
+            $table->decimal('share')->nullable();
             $table->integer('budget')->nullable();
             $table->timestamps();
         });
